// tools/forecast/language/src/language/generated/ast.ts
export type AbstractDeclaration = ApiModel | Attribute | DataModel | DataSource | Enum | FunctionDecl | GeneratorDecl | Input | Interface | OperationGroup | Plugin;

export const AbstractDeclaration = 'AbstractDeclaration';

export function isAbstractDeclaration(item: unknown): item is AbstractDeclaration {
    return reflection.isInstance(item, AbstractDeclaration);
}

export type AttributeName = DataModelAttributeName | DataModelFieldAttributeName | InternalAttributeName;

export function isAttributeName(item: unknown): item is AttributeName {
    return isDataModelAttributeName(item) || isDataModelFieldAttributeName(item) || isInternalAttributeName(item);
}

export type Boolean = boolean;
// ...
export type Expression = ArrayExpr | BinaryExpr | InvocationExpr | LiteralExpr | MemberAccessExpr | NullExpr | ObjectExpr | ReferenceExpr | ThisExpr | UnaryExpr;

export const Expression = 'Expression';

export function isExpression(item: unknown): item is Expression {
    return reflection.isInstance(item, Expression);
}

export type ExpressionType = 'Any' | 'Boolean' | 'DateTime' | 'Float' | 'Int' | 'Null' | 'Object' | 'String' | 'Unsupported';
// ...
export type OperationGroupName = 'Mutation' | 'Query' | 'Subscription';
// ...
export type RegularID = 'abstract' | 'attribute' | 'datasource' | 'enum' | 'in' | 'input' | 'interface' | 'model' | 'plugin' | 'sort' | 'type' | string;
// ...
export interface ApiModel extends AstNode {
    readonly $container: Model;
    readonly $type: 'ApiModel';
    attributes: Array<DataModelAttribute>
    comments: Array<string>
    fields: Array<DataModelField>
    implements: Array<Reference<ApiModel>>
    isExtend: boolean
    name: RegularID
    superTypes: Array<Reference<ApiModel>>
}

export const ApiModel = 'ApiModel';

export function isApiModel(item: unknown): item is ApiModel {
    return reflection.isInstance(item, ApiModel);
}

export interface Argument extends AstNode {
    readonly $container: InvocationExpr;
    readonly $type: 'Argument';
    name?: RegularID
    value: Expression
}

export const Argument = 'Argument';

export function isArgument(item: unknown): item is Argument {
    return reflection.isInstance(item, Argument);
}

export interface ArrayExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'ArrayExpr';
    items: Array<Expression>
}

export const ArrayExpr = 'ArrayExpr';

export function isArrayExpr(item: unknown): item is ArrayExpr {
    return reflection.isInstance(item, ArrayExpr);
}

export interface Attribute extends AstNode {
    readonly $container: Model;
    readonly $type: 'Attribute';
    attributes: Array<InternalAttribute>
    name: AttributeName
    params: Array<AttributeParam>
}

export const Attribute = 'Attribute';

export function isAttribute(item: unknown): item is Attribute {
    return reflection.isInstance(item, Attribute);
}

export interface AttributeArg extends AstNode {
    readonly $container: DataModelAttribute | DataModelFieldAttribute | InternalAttribute;
    readonly $type: 'AttributeArg';
    name?: RegularID
    value: Expression
}

export const AttributeArg = 'AttributeArg';

export function isAttributeArg(item: unknown): item is AttributeArg {
    return reflection.isInstance(item, AttributeArg);
}

export interface AttributeParam extends AstNode {
    readonly $container: Attribute;
    readonly $type: 'AttributeParam';
    default: boolean
    name: RegularID
    type: AttributeParamType
}

export const AttributeParam = 'AttributeParam';
// ...
export interface BinaryExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'BinaryExpr';
    left: Expression
    operator: '!' | '!=' | '&&' | '<' | '<=' | '==' | '>' | '>=' | '?' | '^' | 'in' | '||'
    right: Expression
}

export const BinaryExpr = 'BinaryExpr';

export function isBinaryExpr(item: unknown): item is BinaryExpr {
    return reflection.isInstance(item, BinaryExpr);
}

export interface DataModel extends AstNode {
    readonly $container: Model;
    readonly $type: 'DataModel';
    attributes: Array<DataModelAttribute>
    comments: Array<string>
    fields: Array<DataModelField>
    implements: Array<Reference<DataModel>>
    isAbstract: boolean
    isExtend: boolean
    isView: boolean
    name: RegularID
    superTypes: Array<Reference<DataModel>>
}

export const DataModel = 'DataModel';

export function isDataModel(item: unknown): item is DataModel {
    return reflection.isInstance(item, DataModel);
}

export interface DataModelAttribute extends AstNode {
    readonly $container: ApiModel | DataModel | Enum | Input | Interface | OperationGroup;
    readonly $type: 'DataModelAttribute';
    args: Array<AttributeArg>
    decl: Reference<Attribute>
}

export const DataModelAttribute = 'DataModelAttribute';

export function isDataModelAttribute(item: unknown): item is DataModelAttribute {
    return reflection.isInstance(item, DataModelAttribute);
}

export interface DataModelField extends AstNode {
    readonly $container: ApiModel | DataModel | Enum | FunctionDecl | Input | Interface | OperationGroup;
    readonly $type: 'DataModelField';
    attributes: Array<DataModelFieldAttribute>
    comments: Array<string>
    name: RegularID
    type: DataModelFieldType
}

export const DataModelField = 'DataModelField';

export function isDataModelField(item: unknown): item is DataModelField {
    return reflection.isInstance(item, DataModelField);
}

export interface DataModelFieldAttribute extends AstNode {
    readonly $container: DataModelField | EnumField | Operation;
    readonly $type: 'DataModelFieldAttribute';
    args: Array<AttributeArg>
    decl: Reference<Attribute>
}

export const DataModelFieldAttribute = 'DataModelFieldAttribute';
// ...
export const DataModelFieldType = 'DataModelFieldType';

export function isDataModelFieldType(item: unknown): item is DataModelFieldType {
    return reflection.isInstance(item, DataModelFieldType);
}

export interface DataSource extends AstNode {
    readonly $container: Model;
    readonly $type: 'DataSource';
    fields: Array<DataSourceField>
    name: RegularID
}

export const DataSource = 'DataSource';

export function isDataSource(item: unknown): item is DataSource {
    return reflection.isInstance(item, DataSource);
}

export interface DataSourceField extends AstNode {
    readonly $container: DataSource;
    readonly $type: 'DataSourceField';
    name: RegularID
    value: ArrayExpr | InvocationExpr | LiteralExpr
}

export const DataSourceField = 'DataSourceField';

export function isDataSourceField(item: unknown): item is DataSourceField {
    return reflection.isInstance(item, DataSourceField);
}

export interface Enum extends AstNode {
    readonly $container: Model;
    readonly $type: 'Enum';
    attributes: Array<DataModelAttribute>
    comments: Array<string>
    fields: Array<EnumField>
    name: RegularID
}

export const Enum = 'Enum';

export function isEnum(item: unknown): item is Enum {
    return reflection.isInstance(item, Enum);
}

export interface EnumField extends AstNode {
    readonly $container: ApiModel | DataModel | Enum | FunctionDecl | Input | Interface | OperationGroup;
    readonly $type: 'EnumField';
    attributes: Array<DataModelFieldAttribute>
    comments: Array<string>
    name: RegularID
}

export const EnumField = 'EnumField';

export function isEnumField(item: unknown): item is EnumField {
    return reflection.isInstance(item, EnumField);
}

export interface FieldInitializer extends AstNode {
    readonly $container: ObjectExpr;
    readonly $type: 'FieldInitializer';
    name: RegularID
    value: Expression
}

export const FieldInitializer = 'FieldInitializer';

export function isFieldInitializer(item: unknown): item is FieldInitializer {
    return reflection.isInstance(item, FieldInitializer);
}

export interface FunctionDecl extends AstNode {
    readonly $container: Model;
    readonly $type: 'FunctionDecl';
    attributes: Array<InternalAttribute>
    expression?: Expression
    name: RegularID
    params: Array<FunctionParam>
    returnType: FunctionParamType
}

export const FunctionDecl = 'FunctionDecl';

export function isFunctionDecl(item: unknown): item is FunctionDecl {
    return reflection.isInstance(item, FunctionDecl);
}

export interface FunctionParam extends AstNode {
    readonly $container: ApiModel | DataModel | Enum | FunctionDecl | Input | Interface | OperationGroup;
    readonly $type: 'FunctionParam';
    name: RegularID
    optional: boolean
    type: FunctionParamType
}

export const FunctionParam = 'FunctionParam';

export function isFunctionParam(item: unknown): item is FunctionParam {
    return reflection.isInstance(item, FunctionParam);
}

export interface FunctionParamType extends AstNode {
    readonly $container: FunctionDecl | FunctionParam;
    readonly $type: 'FunctionParamType';
    array: boolean
    reference?: Reference<TypeDeclaration>
    type?: ExpressionType
}

export const FunctionParamType = 'FunctionParamType';

export function isFunctionParamType(item: unknown): item is FunctionParamType {
    return reflection.isInstance(item, FunctionParamType);
}

export interface GeneratorDecl extends AstNode {
    readonly $container: Model;
    readonly $type: 'GeneratorDecl';
    fields: Array<GeneratorField>
    name: RegularID
}

export const GeneratorDecl = 'GeneratorDecl';

export function isGeneratorDecl(item: unknown): item is GeneratorDecl {
    return reflection.isInstance(item, GeneratorDecl);
}

export interface GeneratorField extends AstNode {
    readonly $container: GeneratorDecl;
    readonly $type: 'GeneratorField';
    name: RegularID
    value: ArrayExpr | LiteralExpr
}

export const GeneratorField = 'GeneratorField';

export function isGeneratorField(item: unknown): item is GeneratorField {
    return reflection.isInstance(item, GeneratorField);
}

export interface Input extends AstNode {
    readonly $container: Model;
    readonly $type: 'Input';
    attributes: Array<DataModelAttribute>
    comments: Array<string>
    fields: Array<DataModelField>
    name: RegularID
    superTypes: Array<Reference<Input>>
}

export const Input = 'Input';

export function isInput(item: unknown): item is Input {
    return reflection.isInstance(item, Input);
}

export interface Interface extends AstNode {
    readonly $container: Model;
    readonly $type: 'Interface';
    attributes: Array<DataModelAttribute>
    comments: Array<string>
    fields: Array<DataModelField>
    isExtend: boolean
    name: RegularID
    superTypes: Array<Reference<Interface>>
}

export const Interface = 'Interface';

export function isInterface(item: unknown): item is Interface {
    return reflection.isInstance(item, Interface);
}

export interface InternalAttribute extends AstNode {
    readonly $container: Attribute | FunctionDecl;
    readonly $type: 'InternalAttribute';
    args: Array<AttributeArg>
    decl: Reference<Attribute>
}

export const InternalAttribute = 'InternalAttribute';

export function isInternalAttribute(item: unknown): item is InternalAttribute {
    return reflection.isInstance(item, InternalAttribute);
}

export interface InvocationExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'InvocationExpr';
    args: Array<Argument>
    function: Reference<FunctionDecl>
}

export const InvocationExpr = 'InvocationExpr';

export function isInvocationExpr(item: unknown): item is InvocationExpr {
    return reflection.isInstance(item, InvocationExpr);
}

export interface LiteralExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'LiteralExpr';
    value: Boolean | number | string
}

export const LiteralExpr = 'LiteralExpr';

export function isLiteralExpr(item: unknown): item is LiteralExpr {
    return reflection.isInstance(item, LiteralExpr);
}

export interface MemberAccessExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'MemberAccessExpr';
    member: Reference<DataModelField>
    operand: Expression
}

export const MemberAccessExpr = 'MemberAccessExpr';

export function isMemberAccessExpr(item: unknown): item is MemberAccessExpr {
    return reflection.isInstance(item, MemberAccessExpr);
}

export interface Model extends AstNode {
    readonly $type: 'Model';
    declarations: Array<AbstractDeclaration>
    imports: Array<ModelImport>
}

export const Model = 'Model';
// ...
export interface ObjectExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'ObjectExpr';
    fields: Array<FieldInitializer>
}

export const ObjectExpr = 'ObjectExpr';

export function isObjectExpr(item: unknown): item is ObjectExpr {
    return reflection.isInstance(item, ObjectExpr);
}

export interface Operation extends AstNode {
    readonly $container: ApiModel | DataModel | Enum | FunctionDecl | Input | Interface | OperationGroup;
    readonly $type: 'Operation';
    attributes: Array<DataModelFieldAttribute>
    comments: Array<string>
    name: RegularID
    params: Array<OperationInputParam>
    resultType: DataModelFieldType
}

export const Operation = 'Operation';

export function isOperation(item: unknown): item is Operation {
    return reflection.isInstance(item, Operation);
}

export interface OperationGroup extends AstNode {
    readonly $container: Model;
    readonly $type: 'OperationGroup';
    attributes: Array<DataModelAttribute>
    comments: Array<string>
    fields: Array<Operation>
    isExtend: boolean
    name: OperationGroupName
    superTypes: Array<Reference<Operation>>
}

export const OperationGroup = 'OperationGroup';

export function isOperationGroup(item: unknown): item is OperationGroup {
    return reflection.isInstance(item, OperationGroup);
}

export interface OperationInputParam extends AstNode {
    readonly $container: Operation;
    readonly $type: 'OperationInputParam';
    default: boolean
    defaultValue?: ArrayExpr | LiteralExpr | ObjectExpr
    name: RegularID
    type: DataModelFieldType
}

export const OperationInputParam = 'OperationInputParam';

export function isOperationInputParam(item: unknown): item is OperationInputParam {
    return reflection.isInstance(item, OperationInputParam);
}

export interface Plugin extends AstNode {
    readonly $container: Model;
    readonly $type: 'Plugin';
    fields: Array<PluginField>
    name: RegularID
}

export const Plugin = 'Plugin';

export function isPlugin(item: unknown): item is Plugin {
    return reflection.isInstance(item, Plugin);
}

export interface PluginField extends AstNode {
    readonly $container: Plugin;
    readonly $type: 'PluginField';
    name: RegularID
    value: ArrayExpr | LiteralExpr | ObjectExpr
}

export const PluginField = 'PluginField';
// ...
export interface UnaryExpr extends AstNode {
    readonly $container: Argument | ArrayExpr | AttributeArg | BinaryExpr | DataSourceField | FieldInitializer | FunctionDecl | GeneratorField | MemberAccessExpr | OperationInputParam | PluginField | UnaryExpr | UnsupportedFieldType;
    readonly $type: 'UnaryExpr';
    operand: Expression
    operator: '!'
}

export const UnaryExpr = 'UnaryExpr';

export function isUnaryExpr(item: unknown): item is UnaryExpr {
    return reflection.isInstance(item, UnaryExpr);
}

export interface UnsupportedFieldType extends AstNode {
    readonly $container: DataModelFieldType;
    readonly $type: 'UnsupportedFieldType';
    value: LiteralExpr
}

export const UnsupportedFieldType = 'UnsupportedFieldType';